+++ src/main/java/com/adrian/bank/exchange/CurrencyRateResponse.java
@@ -2,21 +2,13 @@ package com.adrian.bank.exchange;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Map;
+
 @Data // pravi avtomatichni constructori, getteri, setteri
 public class CurrencyRateResponse {
 
-    Map<String, Double> data;
+    Map<String, BigDecimal> data;
 
-//    public CurrencyRateResponse() {
-//    }
-//
-//    public Map<String, Double> getData() {
-//        return data;
-//    }
-//
-//    public void setData(Map<String, Double> data) {
-//        this.data = data;
-//    }
 
 }
